atomic_habits/requirements.txt
django==4.2.17
djangorestframework==3.15.2
Pillow==11.0.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-cors-headers==4.6.0
requests==2.32.3
python-dotenv==1.0.1
celery==5.4.0
kombu==5.4.2
redis==5.2.1
psycopg2==2.9.9
msgpack==1.1.0